requirements.txt
docx==0.2.4
gTTS==2.5.3
nltk==3.9.1
Pillow==10.0.0
Pillow==11.0.0
protobuf==5.29.1
pygame==2.6.1
PyPDF2==3.0.1
python_docx==1.1.2
SpeechRecognition==3.10.4
streamlit==1.39.0
translate==3.6.1
